+++ src/Context/index.jsx
@@ -1,4 +1,4 @@
-import { createContext, useState } from "react";
+import { createContext, useState, useEffect } from "react";
 
 // Creación del contexto
 export const ShoppingCartContext = createContext();
@@ -6,7 +6,7 @@ export const ShoppingCartContext = createContext();
 // Proveedor del contexto
 export const ShoppingCartProvider = ({ children }) => {
   // Shopping Cart · Increment Quantity
-  const [count, setCount] = useState(0);  
+  const [count, setCount] = useState(0);
 
   // Product details · Open/Close
   const [isProductDetailsOpen, setIsProductDetailsOpen] = useState(false);
@@ -27,9 +27,45 @@ export const ShoppingCartProvider = ({ children }) => {
   // Shopping Cart · Order
   const [order, setOrder] = useState([]);
 
+  // Get products · Fetch products
+  const [products, setProducts] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+  
+  // Get products · Search products by title
+  const [searchByTitle, setSearchByTitle] = useState([]);
+  console.log("searchByTitle: ", searchByTitle);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await fetch('https://api.escuelajs.co/api/v1/products');
+        //https://fakestoreapi.com/products
+
+        if (!response.ok) {
+          throw new Error('Error en la solicitud');
+        }
+
+        const jsonData = await response.json();
+        setProducts(jsonData);
+      } catch (err) {
+        setError(err.message);
+        console.error("Error fetching products:", err);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchData();
+  }, []); // El array vacío significa que se ejecuta solo al montar el componente
+
+  /*
+  if (loading) return <div>Cargando...</div>;
+  if (error) return <div>Error: {error}</div>;
+  */
   return (
     <ShoppingCartContext.Provider value={{
-      count, 
+      count,
       setCount,
       isProductDetailsOpen,
       openProductDetails,
@@ -42,7 +78,15 @@ export const ShoppingCartProvider = ({ children }) => {
       openCheckoutSideMenu,
       closeCheckoutSideMenu,
       order,
-      setOrder
+      setOrder,
+      products,
+      setProducts,
+      loading,
+      setLoading,
+      error,
+      setError,
+      searchByTitle,
+      setSearchByTitle
     }}>
       {children}
     </ShoppingCartContext.Provider>
